requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

aiofiles==23.2.1
    # via gradio
aiohttp==3.9.5
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   gradio
    #   httpx
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via aiohttp
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
cn-clip==1.5.1
    # via -r requirements.in
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==2.20.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
faiss-cpu==1.8.0.post1
    # via -r requirements.in
fastapi==0.111.1
    # via gradio
fastapi-cli==0.0.4
    # via fastapi
ffmpy==0.3.2
    # via gradio
filelock==3.15.4
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   triton
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.5.0
    # via
    #   datasets
    #   gradio-client
    #   huggingface-hub
    #   torch
gradio==4.39.0
    # via -r requirements.in
gradio-client==1.1.1
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
huggingface-hub==0.24.2
    # via
    #   datasets
    #   gradio
    #   gradio-client
    #   timm
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.4.0
    # via gradio
jinja2==3.1.4
    # via
    #   fastapi
    #   gradio
    #   torch
kiwisolver==1.4.5
    # via matplotlib
lmdb==1.3.0
    # via cn-clip
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
matplotlib==3.9.1
    # via
    #   -r requirements.in
    #   gradio
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   cn-clip
    #   contourpy
    #   datasets
    #   faiss-cpu
    #   gradio
    #   matplotlib
    #   pandas
    #   pyarrow
    #   torchvision
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.5.82
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
orjson==3.10.6
    # via gradio
packaging==24.1
    # via
    #   datasets
    #   faiss-cpu
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   matplotlib
pandas==2.2.2
    # via
    #   datasets
    #   gradio
pillow==10.4.0
    # via
    #   -r requirements.in
    #   gradio
    #   matplotlib
    #   torchvision
pyarrow==17.0.0
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pydantic==2.8.2
    # via
    #   fastapi
    #   gradio
pydantic-core==2.20.1
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via
    #   fastapi
    #   gradio
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   gradio
    #   huggingface-hub
    #   timm
    #   uvicorn
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
rich==13.7.1
    # via typer
ruff==0.5.5
    # via gradio
safetensors==0.4.3
    # via timm
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   cn-clip
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
sympy==1.13.1
    # via torch
timm==1.0.7
    # via cn-clip
tomlkit==0.12.0
    # via gradio
torch==2.4.0
    # via
    #   -r requirements.in
    #   cn-clip
    #   timm
    #   torchvision
torchvision==0.19.0
    # via
    #   cn-clip
    #   timm
tqdm==4.66.4
    # via
    #   -r requirements.in
    #   cn-clip
    #   datasets
    #   huggingface-hub
triton==3.0.0
    # via torch
typer==0.12.3
    # via
    #   fastapi-cli
    #   gradio
typing-extensions==4.12.2
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   gradio
    #   requests
uvicorn[standard]==0.30.3
    # via
    #   fastapi
    #   gradio
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==11.0.3
    # via
    #   gradio-client
    #   uvicorn
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
